requirements.txt
# Development Dependencies
# Install with: pip install -r requirements-dev.txt

# Code formatting
black>=23.0.0
isort>=5.12.0

# Linting
flake8>=6.0.0
flake8-docstrings>=1.7.0
pylint>=2.17.0

# Type checking
mypy>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0

# Pre-commit hooks
pre-commit>=3.0.0

# PyTorch (CPU version for development)
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pyhessian
curvlinops-for-pytorch
backpack-for-pytorch
